+++ src/errors/FieldFormatError.js
@@ -22,20 +22,15 @@
  * SOFTWARE.
  */
 
-export const ERROR_FIELD_ALLOWED = 'field-allowed';
-export const ERROR_FIELD_DENIED = 'field-denied';
-export const ERROR_FIELD_FORMAT = 'field-format';
-export const ERROR_FIELD_INVALID = 'field-invalid';
-export const ERROR_FIELD_LENGTH = 'field-length';
-export const ERROR_FIELD_MAX = 'field-max';
-export const ERROR_FIELD_MAX_LENGTH = 'field-max-length';
-export const ERROR_FIELD_MAX_WORDS = 'field-max-words';
-export const ERROR_FIELD_MIN = 'field-min';
-export const ERROR_FIELD_MIN_LENGTH = 'field-min-length';
-export const ERROR_FIELD_MIN_WORDS = 'field-min-words';
-export const ERROR_FIELD_NULLABLE = 'field-nullable';
-export const ERROR_FIELD_PATTERN = 'field-pattern';
-export const ERROR_FIELD_REQUIRED = 'field-required';
-export const ERROR_FIELD_TYPE = 'field-type';
-export const ERROR_FIELD_UNKNOWN = 'field-unknown';
-export const ERROR_VALIDATION = 'object-invalid';
+import { ERROR_FIELD_FORMAT } from '../errors';
+import FieldError from './FieldError';
+
+class FieldFormatError extends FieldError {
+  constructor(field, format, path) {
+    super(field, path, ERROR_FIELD_FORMAT);
+    this.format = format;
+    this.message = `"${field}" is not a valid "${format}" format.`;
+  }
+}
+
+export default FieldFormatError;